refactor: abstract the boilerplate code from the consensus unit test
